refactor(embeddings): extract embeddings-pages module
