+++ src/utils/embeddings-pages.ts
@@ -1,0 +1,101 @@
+/**
+ * Page-level embedding: discover retrievable pages on disk, build their
+ * embedding text, embed the changed/cold ones, and merge results into the
+ * existing entry set. Phase 2 swaps the per-page loop for batched embedding.
+ */
+
+import { readdir } from "fs/promises";
+import path from "path";
+import { getProvider } from "./provider.js";
+import { safeReadFile, parseFrontmatter } from "./markdown.js";
+import { CONCEPTS_DIR, QUERIES_DIR } from "./constants.js";
+import { type EmbeddingEntry } from "./embeddings-store.js";
+
+/** A retrievable page record on disk (concepts/ or queries/). */
+export interface PageRecord {
+  slug: string;
+  title: string;
+  summary: string;
+  body: string;
+}
+
+/** Scan concepts/ and queries/ directories, returning retrievable pages. */
+export async function collectPageRecords(root: string): Promise<PageRecord[]> {
+  const records: PageRecord[] = [];
+  for (const dir of [CONCEPTS_DIR, QUERIES_DIR]) {
+    const absDir = path.join(root, dir);
+    let files: string[];
+    try {
+      files = await readdir(absDir);
+    } catch {
+      continue;
+    }
+    for (const file of files.filter((f) => f.endsWith(".md"))) {
+      const record = await readPageRecord(absDir, file);
+      if (record) records.push(record);
+    }
+  }
+  return records;
+}
+
+/** Parse a single page file into a PageRecord, skipping orphans/untitled pages. */
+async function readPageRecord(absDir: string, file: string): Promise<PageRecord | null> {
+  const content = await safeReadFile(path.join(absDir, file));
+  const { meta, body } = parseFrontmatter(content);
+  if (meta.orphaned || typeof meta.title !== "string") return null;
+  return {
+    slug: file.replace(/\.md$/, ""),
+    title: meta.title,
+    summary: typeof meta.summary === "string" ? meta.summary : "",
+    body,
+  };
+}
+
+/** Build the text that represents a page in the embedding space. */
+function buildEmbeddingText(record: PageRecord): string {
+  return record.summary
+    ? `${record.title}\n\n${record.summary}`
+    : record.title;
+}
+
+/**
+ * Embed every page in `records` whose slug appears in `slugsToEmbed`,
+ * returning the new entries. Failures bubble up to the caller.
+ */
+export async function embedPages(
+  records: PageRecord[],
+  slugsToEmbed: Set<string>,
+): Promise<EmbeddingEntry[]> {
+  const provider = getProvider();
+  const now = new Date().toISOString();
+  const fresh: EmbeddingEntry[] = [];
+
+  for (const record of records) {
+    if (!slugsToEmbed.has(record.slug)) continue;
+    const vector = await provider.embed(buildEmbeddingText(record));
+    fresh.push({
+      slug: record.slug,
+      title: record.title,
+      summary: record.summary,
+      vector,
+      updatedAt: now,
+    });
+  }
+  return fresh;
+}
+
+/** Merge fresh embeddings into an existing store, dropping slugs not in liveSlugs. */
+export function mergeEntries(
+  existing: EmbeddingEntry[],
+  fresh: EmbeddingEntry[],
+  liveSlugs: Set<string>,
+): EmbeddingEntry[] {
+  const bySlug = new Map<string, EmbeddingEntry>();
+  for (const entry of existing) {
+    if (liveSlugs.has(entry.slug)) bySlug.set(entry.slug, entry);
+  }
+  for (const entry of fresh) {
+    bySlug.set(entry.slug, entry);
+  }
+  return Array.from(bySlug.values());
+}
