+++ src/utils/embeddings.ts
@@ -16,14 +16,7 @@
  *     and reranking before final page selection.
  */
 
-import { readdir } from "fs/promises";
-import path from "path";
 import { getProvider } from "./provider.js";
-import { safeReadFile, parseFrontmatter } from "./markdown.js";
-import {
-  CONCEPTS_DIR,
-  QUERIES_DIR,
-} from "./constants.js";
 import { hashChunkText, splitIntoChunks } from "./retrieval.js";
 import * as output from "./output.js";
 import {
@@ -36,95 +29,12 @@ import {
   isStoreEmpty,
   resolveEmbeddingModel,
 } from "./embeddings-store.js";
-
-/** A retrievable page record on disk (concepts/ or queries/). */
-interface PageRecord {
-  slug: string;
-  title: string;
-  summary: string;
-  body: string;
-}
-
-/** Scan concepts/ and queries/ directories, returning retrievable pages. */
-async function collectPageRecords(root: string): Promise<PageRecord[]> {
-  const records: PageRecord[] = [];
-  for (const dir of [CONCEPTS_DIR, QUERIES_DIR]) {
-    const absDir = path.join(root, dir);
-    let files: string[];
-    try {
-      files = await readdir(absDir);
-    } catch {
-      continue;
-    }
-    for (const file of files.filter((f) => f.endsWith(".md"))) {
-      const record = await readPageRecord(absDir, file);
-      if (record) records.push(record);
-    }
-  }
-  return records;
-}
-
-/** Parse a single page file into a PageRecord, skipping orphans/untitled pages. */
-async function readPageRecord(absDir: string, file: string): Promise<PageRecord | null> {
-  const content = await safeReadFile(path.join(absDir, file));
-  const { meta, body } = parseFrontmatter(content);
-  if (meta.orphaned || typeof meta.title !== "string") return null;
-  return {
-    slug: file.replace(/\.md$/, ""),
-    title: meta.title,
-    summary: typeof meta.summary === "string" ? meta.summary : "",
-    body,
-  };
-}
-
-/** Build the text that represents a page in the embedding space. */
-function buildEmbeddingText(record: PageRecord): string {
-  return record.summary
-    ? `${record.title}\n\n${record.summary}`
-    : record.title;
-}
-
-/**
- * Embed every page in `records` whose slug appears in `slugsToEmbed`,
- * returning the new entries. Failures bubble up to the caller.
- */
-async function embedPages(
-  records: PageRecord[],
-  slugsToEmbed: Set<string>,
-): Promise<EmbeddingEntry[]> {
-  const provider = getProvider();
-  const now = new Date().toISOString();
-  const fresh: EmbeddingEntry[] = [];
-
-  for (const record of records) {
-    if (!slugsToEmbed.has(record.slug)) continue;
-    const vector = await provider.embed(buildEmbeddingText(record));
-    fresh.push({
-      slug: record.slug,
-      title: record.title,
-      summary: record.summary,
-      vector,
-      updatedAt: now,
-    });
-  }
-  return fresh;
-}
-
-/** Merge fresh embeddings into an existing store, dropping slugs not in liveSlugs. */
-function mergeEntries(
-  existing: EmbeddingEntry[],
-  fresh: EmbeddingEntry[],
-  liveSlugs: Set<string>,
-): EmbeddingEntry[] {
-  const bySlug = new Map<string, EmbeddingEntry>();
-  for (const entry of existing) {
-    if (liveSlugs.has(entry.slug)) bySlug.set(entry.slug, entry);
-  }
-  for (const entry of fresh) {
-    bySlug.set(entry.slug, entry);
-  }
-  return Array.from(bySlug.values());
-}
+import {
+  type PageRecord,
+  collectPageRecords,
+  embedPages,
+  mergeEntries,
+} from "./embeddings-pages.js";
 
 /**
  * Refresh chunk embeddings for the given pages, reusing existing chunk vectors
